logger updates

+++ xlive/Blam/Engine/render/render_section_visibility.h
@@ -3,9 +3,9 @@
 void __cdecl render_section_visibility_compute(
 	int32 a1,
 	int32 a2,
-	uint8 model_group_index,
-	int32 a4,
-	int32 a5,
-	int32 a6,
-	int32 a7,
+	int32 model_group_index,
+	uint32 a4,
+	uint32 a5,
+	uint32 a6,
+	uint32 a7,
 	bool a8);
